Some fix

+++ project_2/utils/ber.m
@@ -1,7 +1,7 @@
 function [ p ] = ber( y, y_hat )
 %BER Summary of this function goes here
 %   Detailed explanation goes here
-c = confusionmat(y, y_hat);
+c = confusionmat(double(y), double(y_hat));
 num_class = size(unique(y), 1);
 p = 0;
 for i=1:num_class
@@ -10,4 +10,3 @@ end
 p = p/num_class;
 
 end
-
